Silencing a build warning.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@5923 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/simulations/current/makeModelSlice/MakeAllModelSlicesApp.cc
@@ -70,7 +70,7 @@ int MakeAllModelSlicesApp::run(int argc, char* argv[])
 	std::string slicebase = subset.getString("slicename");
 
 
-	for (int chan=comms.rank()*sliceWidth; chan < nchan; chan += comms.nProcs()*sliceWidth) {
+	for (size_t chan=comms.rank()*sliceWidth; chan < nchan; chan += comms.nProcs()*sliceWidth) {
 
 	    std::stringstream slicename;
 	    slicename << slicebase << chan/sliceWidth;
